chg: usr: Scripts automatically install with Python package.
Added generator function yield_rows() to all abstractions.
